requirements.txt
bleach==1.4.2
blessed==1.9.5
boto==2.39.0
botocore==1.4.12
cement==2.8.2
colorama==0.3.7
Django==1.9.2
django-appconf==1.0.2
django-cors-headers==1.1.0
django-crispy-forms==1.6.0
django-floppyforms==1.6.2
django-imagekit==3.3
django-markdownify==0.1.0
django-rest-auth==0.7.0
django-simple-history==1.8.0
django-storages-redux==1.3.2
djangorestframework==3.3.2
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
html5lib==0.9999999
jmespath==0.9.0
Markdown==2.6.5
mysqlclient==1.3.7
pathspec==0.3.4
pilkit==1.1.13
Pillow==3.2.0
python-dateutil==2.5.2
pytz==2016.4
PyYAML==3.11
requests==2.9.1
semantic-version==2.5.0
six==1.10.0
texttable==0.8.4
wcwidth==0.1.6
websocket-client==0.37.0
